Make helpers into concrete functions for WASM binary size purposes
