+++ leptos_dom/src/macro_helpers/into_property.rs
@@ -77,6 +77,38 @@ prop_type!(f64);
 prop_type!(bool);
 
 #[cfg(all(target_arch = "wasm32", feature = "web"))]
+use std::borrow::Cow;
+
+#[cfg(all(target_arch = "wasm32", feature = "web"))]
+pub(crate) fn property_helper(
+    el: &web_sys::Element,
+    name: Cow<'static, str>,
+    value: Property
+) {
+    use leptos_reactive::create_render_effect;
+
+      match value {
+        Property::Fn(cx, f) => {
+          let el = el.clone();
+          create_render_effect(cx, move |old| {
+            let new = f();
+            let prop_name = wasm_bindgen::intern(&name);
+            if old.as_ref() != Some(&new)
+              && !(old.is_none() && new == wasm_bindgen::JsValue::UNDEFINED)
+            {
+              property_expression(&el, prop_name, new.clone())
+            }
+            new
+          });
+        }
+        Property::Value(value) => {
+          let prop_name = wasm_bindgen::intern(&name);
+          property_expression(el, prop_name, value)
+        }
+      };
+}
+
+#[cfg(all(target_arch = "wasm32", feature = "web"))]
 pub(crate) fn property_expression(
   el: &web_sys::Element,
   prop_name: &str,
